Graveyard/requirements.txt
asgiref==3.8.1
Django==5.1.6
mysqlclient==2.2.7
pycountry==24.6.1
sqlparse==0.5.3
typing_extensions==4.12.2
gunicorn==20.1.0
djangorestframework==3.16.0
